function: ListViewをnavigatorによる遷移にあわせて機能分割
