+++ package.json
@@ -14,6 +14,7 @@
     "@react-navigation/bottom-tabs": "^5.11.11",
     "@react-navigation/material-bottom-tabs": "^5.3.15",
     "@react-navigation/native": "^5.9.4",
+    "@react-navigation/stack": "^5.14.5",
     "dayjs": "^1.10.5",
     "expo": "~41.0.1",
     "expo-file-system": "^11.1.3",
